Even more work to CMS project

+++ lesson_8/CMS_project/cms.rb
@@ -2,11 +2,14 @@ require "sinatra"
 require "sinatra/reloader"
 require "tilt/erubis"
 require "redcarpet"
+require "yaml"
 
 configure do
   enable :sessions
   set :session_secret, SecureRandom.hex(32)
 end
+
+# Helper Methods
 
 def data_path
   if ENV["RACK_ENV"] == "test"
@@ -32,27 +35,60 @@ def load_file_content(path)
   end
 end
 
+def user_signed_in?
+  session.key?(:username)
+end
+
+def require_signed_in_user
+  unless user_signed_in?
+    session[:message] = "You must be signed in to do that."
+    redirect '/'
+  end
+end
+
+def load_user_credentials
+  credentials_path = if ENV["RACK_ENV"] == "test"
+    File.expand_path("../test/users.yml", __FILE__)
+  else
+    File.expand_path("../users.yml", __FILE__)
+  end
+
+  YAML.load_file(credentials_path)
+end
+
+def valid_user?(username, password)
+  users = load_user_credentials
+  users[username] == password
+end
+
+# Requests
+
 get "/" do
   pattern = File.join(data_path, "*")
   @files = Dir.glob(pattern).map do |path|
     File.basename(path)
   end
-  erb :index, layout: :layout
+  erb :index
+end
+
+get "/test" do
+  users = YAML.load_file(File.expand_path('../users.yml', __FILE__))
+  users.keys
 end
 
 get "/users/signin" do
-  erb :signin, layout: :layout
+  erb :signin
 end
 
 post "/users/signin" do
-  if params[:username] == 'admin' && params[:password] == 'secret'
+  if valid_user?(params[:username], params[:password])
     session[:username] = params[:username]    
     session[:message] = 'Welcome!'
     redirect '/'
   else
     session[:message] = 'Invalid credentials.'
     status 422
-    erb :signin, layout: :layout
+    erb :signin
   end
 end
 
@@ -63,10 +99,17 @@ post "/users/signout" do
 end
 
 get '/new' do
+  unless session[:username]
+    session[:message] = "You must be signed in to do that."
+    redirect '/'
+  end
+
   erb :new, layout: :layout
 end
 
 post '/create' do
+  user_signed_in?
+
   filename = params[:filename].to_s
 
   if filename.size == 0
@@ -98,15 +141,19 @@ get "/:filename" do
 end
 
 get "/:filename/edit" do
+  user_signed_in?
+
   file_path = File.join(data_path, params[:filename])
 
   @filename = params[:filename]
   @content = File.read(file_path)
 
-  erb :edit, layout: :layout
+  erb :edit
 end
 
 post "/:filename" do
+  user_signed_in?
+
   file_path = File.join(data_path, params[:filename])
 
   File.write(file_path, params[:content])
@@ -116,6 +163,11 @@ post "/:filename" do
 end
 
 post "/:filename/delete" do
+  unless session[:username]
+    session[:message] = "You must be signed in to do that."
+    redirect '/'
+  end
+
   file_path = File.join(data_path, params[:filename])
 
   File.delete(file_path)
@@ -123,3 +175,5 @@ post "/:filename/delete" do
   session[:message] = "'#{params[:filename]}' has been deleted."
   redirect "/"
 end
+
+
